move injest data to results
